+++ src/queue.hh
@@ -59,7 +59,7 @@ class Queue {
             return type_;
         }
 
-        std::pair<Node*, tracking_vector<unsigned short, DataStruct::Tree> > select(CacheTree* tree, VECTOR captured, size_t thread_id) {
+        std::pair<Node*, tracking_vector<unsigned short, DataStruct::Tree> > select(CacheTree* tree, VECTOR captured, unsigned short thread_id) {
 begin:
             int cnt;
             tracking_vector<unsigned short, DataStruct::Tree> prefix;
@@ -115,7 +115,7 @@ begin:
 };
 
 extern int bbound(CacheTree* tree, size_t max_num_nodes, Queue* q, PermutationMap* p,
-    size_t thread_id, double* min_objective);
+    unsigned short thread_id, double* min_objective);
 
 extern void bbound_init(CacheTree* tree);
